examples/080-pipecat-voice-pipeline-python/requirements.txt
pipecat-ai[deepgram,daily,openai,silero]==0.0.108
deepgram-sdk==6.1.1
python-dotenv==1.2.2
